Map bindings for `REDISMODULE_NOTIFY_*` to the correct type

+++ src/raw.rs
@@ -88,17 +88,17 @@ impl From<Status> for Result<(), &str> {
 
 #[cfg(feature = "experimental-api")]
 bitflags! {
-    pub struct NotifyEvent : i32 {
-        const GENERIC = REDISMODULE_NOTIFY_GENERIC as i32;
-        const STRING = REDISMODULE_NOTIFY_STRING as i32;
-        const LIST = REDISMODULE_NOTIFY_LIST as i32;
-        const SET = REDISMODULE_NOTIFY_SET as i32;
-        const HASH = REDISMODULE_NOTIFY_HASH as i32;
-        const ZSET = REDISMODULE_NOTIFY_ZSET as i32;
-        const EXPIRED = REDISMODULE_NOTIFY_EXPIRED as i32;
-        const EVICTED = REDISMODULE_NOTIFY_EVICTED as i32;
-        const STREAM = REDISMODULE_NOTIFY_STREAM as i32;
-        const ALL = REDISMODULE_NOTIFY_ALL as i32;
+    pub struct NotifyEvent : c_int {
+        const GENERIC = REDISMODULE_NOTIFY_GENERIC;
+        const STRING = REDISMODULE_NOTIFY_STRING;
+        const LIST = REDISMODULE_NOTIFY_LIST;
+        const SET = REDISMODULE_NOTIFY_SET;
+        const HASH = REDISMODULE_NOTIFY_HASH;
+        const ZSET = REDISMODULE_NOTIFY_ZSET;
+        const EXPIRED = REDISMODULE_NOTIFY_EXPIRED;
+        const EVICTED = REDISMODULE_NOTIFY_EVICTED;
+        const STREAM = REDISMODULE_NOTIFY_STREAM;
+        const ALL = REDISMODULE_NOTIFY_ALL;
     }
 }
